Split main.rs into modules by concern
